cdeploy: improved simulation mode
when simulating the deployment of a file and the target directory does not exist, the whole file system hierarchy is going to be tested in reverse order until the first existing directory is reached. This directory is then tested for write permission. The result of this test determines whether the substructure could be created and the file then be deployed.